feat: 添加job local file report

+++ core/src/main/java/com/jdragon/aggregation/core/taskgroup/runner/AbstractRunner.java
@@ -15,7 +15,7 @@ public class AbstractRunner implements Runnable {
 
     private Configuration jobConf;
 
-    private int jobId;
+    private long jobId;
 
     private Communication runnerCommunication;
